refactor: move isTruthy / isFalsy to utils/bool.ts
