+++ src/unsafe.ts
@@ -1,15 +1,5 @@
 import { ParseError } from "./error.ts";
-
-const TRUTHY = ["true", "1", "yes", "y", "on"];
-const FALSY = ["false", "0", "no", "n", "off"];
-
-export function isTruthy(value: string): boolean {
-	return TRUTHY.includes(value);
-}
-
-export function isFalsy(value: string): boolean {
-	return FALSY.includes(value);
-}
+import { isFalsy, isTruthy } from "./utils/bool.ts";
 
 /**
  * A parser for environment variables that attempts to parse the string
